requirements.txt
pytest==7.3.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
selenium==4.9.0
xlrd==2.0.1
webdriver-manager==3.8.6
